remove circular import of IfVisible -> Timer -> IfVisible
